requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
annotated-types==0.6.0
antlr4-python3-runtime==4.9.3
anyio==3.7.1
apiclient==1.0.4
asgiref==3.2.10
async-timeout==4.0.3
attrs==23.1.0
backoff==2.2.1
blinker==1.6.2
botocore==1.31.52
CacheControl==0.13.1
cachetools==4.1.1
certifi==2020.6.20
cffi==1.15.1
chardet==3.0.4
charset-normalizer==3.2.0
clarifai==9.9.3
clarifai-grpc==9.8.1
click==7.1.2
cmake==3.27.0
cohere==4.32
colorama==0.4.6
comtypes==1.2.0
contextlib2==21.6.0
contourpy==1.1.1
cryptography==41.0.3
cvzone==1.5.6
cycler==0.12.1
dataclasses-json==0.6.1
dill==0.3.7
distlib==0.3.7
Django==3.1.2
django-rest-framework==0.1.0
djangorestframework==3.14.0
dlib==19.24.2
docopt==0.6.2
face-recognition==1.3.0
face-recognition-models==0.3.0
fastavro==1.8.2
filelock==3.12.2
firebase-admin==6.2.0
Flask==1.1.2
Flask-Login==0.6.2
Flask-SQLAlchemy==3.0.5
fonttools==4.43.1
frozenlist==1.4.0
fsspec==2023.10.0
google-api-core==2.12.0
google-api-python-client==1.12.5
google-auth==2.23.3
google-auth-httplib2==0.0.4
google-cloud-aiplatform==1.35.0
google-cloud-bigquery==3.12.0
google-cloud-core==2.3.3
google-cloud-firestore==2.11.1
google-cloud-resource-manager==1.10.4
google-cloud-storage==2.10.0
google-cloud-texttospeech==2.14.2
google-crc32c==1.5.0
google-resumable-media==2.5.0
googleapis-common-protos==1.61.0
gpt4all==2.0.1
greenlet==2.0.2
grpc-google-iam-v1==0.12.6
grpcio==1.56.2
grpcio-status==1.56.2
gunicorn==21.2.0
httplib2==0.18.1
huggingface-hub==0.17.3
idna==2.10
imageio==2.31.5
importlib-metadata==6.8.0
iniconfig==2.0.0
itsdangerous==1.1.0
Jinja2==2.11.2
jmespath==1.0.1
joblib==1.3.2
jsonpatch==1.33
jsonpointer==2.4
kiwisolver==1.4.5
langchain==0.0.323
langsmith==0.0.52
lazy_loader==0.3
llvmlite==0.41.1
lxml==4.9.3
manifest-ml==0.0.1
markdown-it-py==3.0.0
MarkupSafe==1.1.1
marshmallow==3.20.1
matplotlib==3.8.0
mdurl==0.1.2
more-itertools==10.1.0
mpmath==1.3.0
msgpack==1.0.5
multidict==6.0.4
mypy-extensions==1.0.0
networkx==3.1
nlpcloud==1.1.44
nltk==3.6.2
numba==0.58.1
numpy==1.22.0
omegaconf==2.2.3
openai==0.28.1
openai-whisper @ git+https://github.com/openai/whisper.git@b38a1f20f4b23f3f3099af2c3e0ca95627276ddf
opencv-python==4.7.0.68
openlm==0.0.5
packaging==23.1
pandas==1.3.5
Pillow==10.0.0
pipreqs==0.4.13
platformdirs==3.10.0
pluggy==1.3.0
proto-plus==1.22.3
protobuf==4.24.4
psycopg2-binary==2.9.7
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAudio==0.2.13
pycocotools==2.0.6
pycparser==2.21
pydantic==2.4.2
pydantic_core==2.10.1
Pygments==2.16.1
PyJWT==2.8.0
pyparsing==3.1.1
PyPDF2==3.0.1
pypiwin32==223
pytest==7.4.1
python-dateutil==2.8.2
python-docx==0.8.11
python-dotenv==1.0.0
python-rapidjson==1.12
pyttsx3==2.90
pytz==2020.1
pywin32==306
PyYAML==6.0.1
redis==5.0.1
regex==2023.8.8
requests==2.31.0
rich==13.4.2
rsa==4.6
s3transfer==0.6.2
safetensors==0.4.0
schema==0.7.5
scikit-image==0.22.0
scipy==1.11.3
shapely==2.0.2
six==1.15.0
sniffio==1.3.0
SQLAlchemy==2.0.19
sqlitedict==2.1.0
sqlparse==0.4.1
static3==0.7.0
sympy==1.12
tenacity==8.2.3
tifffile==2023.9.26
tiktoken==0.3.3
tokenizers==0.14.1
torch==2.1.0
tqdm==4.66.1
transformers==4.34.1
tritonclient==2.34.0
typing-inspect==0.9.0
typing_extensions==4.7.1
tzdata==2023.3
uritemplate==3.0.1
urllib3==2.0.7
virtualenv==20.24.2
Werkzeug==1.0.1
whitenoise==6.5.0
wincertstore==0.2
yarg==0.1.9
yarl==1.9.2
zipp==3.17.0
